+++ iree/compiler/Conversion/LinalgToLLVMGPU/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# iree/compiler/Conversion/LinalgToNVVM/BUILD                                  #
+# iree/compiler/Conversion/LinalgToLLVMGPU/BUILD                               #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -12,13 +12,18 @@ iree_add_all_subdirs()
 
 iree_cc_library(
   NAME
-    LinalgToNVVM
+    LinalgToLLVMGPU
   HDRS
+    "Common.h"
     "KernelConfig.h"
+    "LLVMGPUCodeGenOptions.h"
     "Passes.h"
   SRCS
+    "Common.cpp"
     "ConvertToNVVM.cpp"
+    "ConvertToROCDL.cpp"
     "KernelConfig.cpp"
+    "LLVMGPUCodeGenOptions.cpp"
     "Passes.cpp"
     "RemoveTrivialLoops.cpp"
     "TileAndDistribute.cpp"
@@ -27,6 +32,7 @@ iree_cc_library(
     MLIRAffine
     MLIRGPU
     MLIRGPUToNVVMTransforms
+    MLIRGPUToROCDLTransforms
     MLIRIR
     MLIRLinalg
     MLIRLinalgTransforms
