+++ iree/compiler/Conversion/LinalgToLLVMGPU/ConvertToROCDL.cpp
@@ -1,0 +1,114 @@
+// Copyright 2021 Nod Labs
+// Copyright 2021 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "iree/compiler/Conversion/LinalgToLLVMGPU/Passes.h"
+#include "iree/compiler/Conversion/LinalgToLLVMGPU/Common.h"
+#include "iree/compiler/Conversion/CodegenUtils/FunctionUtils.h"
+#include "iree/compiler/Dialect/IREE/IR/IREEOps.h"
+#include "mlir/Conversion/GPUToROCDL/GPUToROCDLPass.h"
+#include "mlir/Conversion/StandardToLLVM/ConvertStandardToLLVM.h"
+#include "mlir/Conversion/VectorToLLVM/ConvertVectorToLLVM.h"
+#include "mlir/Dialect/GPU/Passes.h"
+#include "mlir/Dialect/LLVMIR/ROCDLDialect.h"
+#include "mlir/Dialect/StandardOps/IR/Ops.h"
+#include "mlir/Dialect/Vector/VectorOps.h"
+#include "mlir/Transforms/GreedyPatternRewriteDriver.h"
+#include "mlir/Transforms/Passes.h"
+
+namespace mlir {
+namespace iree_compiler {
+
+namespace {
+
+/// A pass that replaces all occurrences of GPU device operations with their
+/// corresponding ROCDL equivalent.
+///
+/// This pass only handles device code and is not meant to be run on GPU host
+/// code.
+struct ConvertToROCDLPass
+    : public PassWrapper<ConvertToROCDLPass, OperationPass<ModuleOp>> {
+  void getDependentDialects(DialectRegistry &registry) const override {
+    registry.insert<LLVM::LLVMDialect, ROCDL::ROCDLDialect>();
+  }
+  void runOnOperation() override {
+    ModuleOp m = getOperation();
+
+    /// Customize the bitwidth used for the device side index computations.
+    LowerToLLVMOptions options(m.getContext(), DataLayout(m));
+    options.overrideIndexBitwidth(64);
+    LLVMTypeConverter converter(m.getContext(), options);
+    // Apply in-dialect lowering first. In-dialect lowering will replace ops
+    // which need to be lowered further, which is not supported by a single
+    // conversion pass.
+    // Run Vector -> Vector transformations ahead of conversion to LLVM.
+    {
+      OwningRewritePatternList patterns(&getContext());
+      vector::populateVectorToVectorCanonicalizationPatterns(patterns);
+      vector::populateVectorSlicesLoweringPatterns(patterns);
+      vector::populateVectorContractLoweringPatterns(
+          patterns,
+          vector::VectorTransformsOptions().setVectorTransformsOptions(
+              vector::VectorContractLowering::OuterProduct));
+      mlir::vector::populateVectorTransposeLoweringPatterns(patterns);
+      (void)applyPatternsAndFoldGreedily(m, std::move(patterns));
+    }
+    {
+      OwningRewritePatternList patterns(&getContext());
+      populateGpuRewritePatterns(patterns);
+      (void)applyPatternsAndFoldGreedily(m, std::move(patterns));
+    }
+    {
+      OwningRewritePatternList llvmPatterns(&getContext());
+      llvmPatterns.insert<ConvertFunc, ConvertIREEBindingOp>(m.getContext(),
+                                                             converter);
+      llvmPatterns
+          .insert<HALInterfaceWorkgroupOpsConverter<
+                      IREE::HAL::InterfaceWorkgroupIDOp, ROCDL::BlockIdXOp,
+                      ROCDL::BlockIdYOp, ROCDL::BlockIdZOp>,
+                  HALInterfaceWorkgroupOpsConverter<
+                      IREE::HAL::InterfaceWorkgroupCountOp, ROCDL::GridDimXOp,
+                      ROCDL::GridDimYOp, ROCDL::GridDimZOp>,
+                  HALInterfaceWorkgroupOpsConverter<
+                      IREE::HAL::InterfaceWorkgroupSizeOp, ROCDL::BlockDimXOp,
+                      ROCDL::BlockDimYOp, ROCDL::BlockDimZOp>>(m.getContext());
+      populateStdToLLVMConversionPatterns(converter, llvmPatterns);
+      populateVectorToLLVMConversionPatterns(converter, llvmPatterns);
+      populateGpuToROCDLConversionPatterns(converter, llvmPatterns);
+      LLVMConversionTarget target(getContext());
+      populateStdToLLVMFuncOpConversionPattern(converter, llvmPatterns);
+      configureGpuToROCDLConversionLegality(target);
+      target.addDynamicallyLegalOp<FuncOp>([&](FuncOp funcOp) {
+        if (isEntryPoint(funcOp)) return false;
+        return true;
+      });
+      if (failed(applyPartialConversion(m, target, std::move(llvmPatterns))))
+        signalPassFailure();
+    }
+  }
+};
+
+}  // anonymous namespace
+
+std::unique_ptr<OperationPass<ModuleOp>> createConvertToROCDLPass() {
+  return std::make_unique<ConvertToROCDLPass>();
+}
+
+static PassRegistration<ConvertToROCDLPass> pass(
+    "iree-codegen-convert-to-rocdl",
+    "Perform final conversion from builtin/GPU/HAL/standard dialect to LLVM "
+    "and ROCDL dialects");
+
+}  // namespace iree_compiler
+}  // namespace mlir
