+++ iree/compiler/Conversion/LinalgToLLVMGPU/Passes.h
@@ -12,9 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef IREE_COMPILER_CONVERSION_LINALGTONVVM_PASSES_H_
-#define IREE_COMPILER_CONVERSION_LINALGTONVVM_PASSES_H_
+#ifndef IREE_COMPILER_CONVERSION_LINALGTOLLVMGPU_PASSES_H_
+#define IREE_COMPILER_CONVERSION_LINALGTOLLVMGPU_PASSES_H_
 
+#include "iree/compiler/Conversion/LinalgToLLVMGPU/LLVMGPUCodeGenOptions.h"
 #include "iree/compiler/Dialect/HAL/IR/HALOps.h"
 #include "mlir/Pass/Pass.h"
 
@@ -23,6 +24,9 @@ namespace iree_compiler {
 
 /// Performs the final conversion to NNVM+LLVM dialect.
 std::unique_ptr<OperationPass<ModuleOp>> createConvertToNVVMPass();
+
+/// Performs the final conversion to ROCDL+LLVM dialect.
+std::unique_ptr<OperationPass<ModuleOp>> createConvertToROCDLPass();
 
 /// Convert Linalg ops to Vector.
 std::unique_ptr<OperationPass<FuncOp>> createVectorizationPass();
@@ -33,11 +37,12 @@ createTileAndDistributeToThreads();
 
 std::unique_ptr<OperationPass<FuncOp>> createRemoveSingleIterationLoopPass();
 
-/// Populates passes needed to lower a XLA HLO op to NVVM dialect via the
+/// Populates passes needed to lower a XLA HLO op to NVVM/ROCDL dialect via the
 /// structured ops path. The pass manager `pm` in here should operate on the
 /// module within the IREE::HAL::ExecutableOp.
-void buildNVVMTransformPassPipeline(OpPassManager &pm);
+void buildLLVMGPUTransformPassPipeline(OpPassManager &pm, const LLVMGPUCodegenOptions &options);
+
 }  // namespace iree_compiler
 }  // namespace mlir
 
-#endif  // IREE_COMPILER_CONVERSION_LINALGTONVVM_PASSES_H_
+#endif  // IREE_COMPILER_CONVERSION_LINALGTOLLVMGPU_PASSES_H_
