+++ iree/compiler/Conversion/LinalgToLLVMGPU/test/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# iree/compiler/Conversion/LinalgToNVVM/test/BUILD                             #
+# iree/compiler/Conversion/LinalgToLLVMGPU/test/BUILD                          #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -15,9 +15,11 @@ iree_lit_test_suite(
     lit
   SRCS
     "convert_to_nvvm.mlir"
+    "convert_to_rocdl.mlir"
     "distribute_to_thread.mlir"
-    "pipeline_test.mlir"
+    "nvvm_pipeline_test.mlir"
     "remove_loops.mlir"
+    "rocdl_pipeline_test.mlir"
     "vectorization.mlir"
   DATA
     iree::tools::IreeFileCheck
